Update for Issue #19
Made great progress on the issue. It is now possible to transfer the
library of a song, though any edits made afterwards will result in a
crash. Perhaps it has something to do with some piece of code still
remembering the now obsolete object. Will check for that in the morning.